% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carelesswhisper.R
\docType{data}
\name{jfk}
\alias{jfk}
\title{Audio sample for testing}
\format{
An object of class \code{audioSample} of length 176000.
}
\usage{
jfk
}
\description{
Audio sample for testing
}
\keyword{datasets}
