% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carelesswhisper.R
\name{record_audio}
\alias{record_audio}
\title{Record audio from the default input device}
\usage{
record_audio(seconds)
}
\arguments{
\item{seconds}{recording length}
}
\value{
Numeric vector of mono sound data sampled at 16kHz
}
\description{
Record audio from the default input device
}
