% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carelesswhisper.R
\name{whisper}
\alias{whisper}
\title{Perform automatic speech recognition of the given sound sample}
\usage{
whisper(ctx, snd, params = list(), verbose = FALSE, details = FALSE)
}
\arguments{
\item{ctx}{whisper context (which you have previously created using \code{whisper_init()})}

\item{snd}{Sound data.  16kHz mono audio in a numeric vector 
with all values in the range [-1, 1].  This package includes the function
 `record_audio()` which will record audio in this format.
 You could also use \code{audio::record()} or any other audio package
 you have access to.}

\item{params}{parameters for whisper. A user should usually create a default set 
of parameters by calling
 \code{whisper_param_defaults()} and then modify.}

\item{verbose}{logical. be verbose? default: FALSE.}

\item{details}{logical. return detailed breakdown as a data.frame?  default: FALSE}
}
\value{
Character string
}
\description{
Perform automatic speech recognition of the given sound sample
}
\examples{
\dontrun{
  ctx <- whisper_init()  # Initialise the model
  snd <- record_audio(2) # record 2 seconds of audio 
  whisper(ctx, snd)      # perform speech recognition
}

}
