% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carelesswhisper.R
\name{whisper_default_params}
\alias{whisper_default_params}
\title{Fetch a copy of the default whisper parameters}
\usage{
whisper_default_params()
}
\value{
Named list of default parameters
}
\description{
\describe{
   \item{n_threads}{Number of threads to use when processing. Default: 4}
   \item{translate}{Translate from source language into english? Default: FALSE}
   \item{language}{language represented in audio.  Use 'auto' to automatically
         detect language. Default: 'en'}
}
}
