% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carelesswhisper.R
\name{whisper_init}
\alias{whisper_init}
\title{Initialise whisper by loading a model}
\usage{
whisper_init(
  model_path = system.file("ggml-tiny.bin", package = "carelesswhisper", mustWork = TRUE),
  verbose = FALSE
)
}
\arguments{
\item{model_path}{path to whisper.cpp model. By default
this will use the "ggml-tiny.bin" file included with this 
package installation, which is a tiny multi-language model 
See README for this package, or the original 
whisper.cpp documentation, for how to download other models.}

\item{verbose}{Be verbose about model initialisation?  Logical. Default: FALSE}
}
\value{
whisper context (\code{ctx})
}
\description{
Initialise whisper by loading a model
}
