% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carelesswhisper.R
\docType{data}
\name{whisper_lang_codes}
\alias{whisper_lang_codes}
\title{Named list of two-letter language codes to use as \code{language} parameter}
\format{
An object of class \code{list} of length 99.
}
\usage{
whisper_lang_codes
}
\description{
Named list of two-letter language codes to use as \code{language} parameter
}
\keyword{datasets}
