% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miraiwhisper.R
\name{whisper_rt}
\alias{whisper_rt}
\title{Real Time Audio Transcription using Whisper}
\usage{
whisper_rt(seconds = 5L)
}
\arguments{
\item{seconds}{a recording of this length is sampled repeatedly.}
}
\value{
Invisible NULL. The transcribed text is output to the console.
}
\description{
Real Time Audio Transcription using Whisper
}
\details{
Send an interrupt to break and exit from the function.
}
